Refactored controller and packages
