+++ src/main/java/no/fintlabs/datasync/DeltaSyncProducerKafka.java
@@ -1,7 +1,8 @@
-package no.fintlabs.kafka;
+package no.fintlabs.datasync;
 
 import lombok.extern.slf4j.Slf4j;
 import no.fintlabs.adapter.models.SyncPageMetadata;
+import no.fintlabs.kafka.EventProducerKafka;
 import no.fintlabs.kafka.event.EventProducerFactory;
 import no.fintlabs.kafka.event.topic.EventTopicService;
 import org.springframework.stereotype.Service;
