+++ src/main/java/no/fintlabs/heartbeat/HeartbeatKafkaProducer.java
@@ -1,6 +1,7 @@
-package no.fintlabs.kafka;
+package no.fintlabs.heartbeat;
 
 import no.fintlabs.adapter.models.AdapterHeartbeat;
+import no.fintlabs.kafka.EventProducerKafka;
 import no.fintlabs.kafka.event.EventProducerFactory;
 import no.fintlabs.kafka.event.topic.EventTopicService;
 import org.springframework.stereotype.Service;
