+++ src/main/java/no/fintlabs/utils/FintListenableFutureCallback.java
@@ -1,4 +1,4 @@
-package no.fintlabs;
+package no.fintlabs.utils;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.kafka.support.SendResult;
